[Search] Share code between AR and elastic
