+++ app/logical/danbooru/paginator/elasticsearch_extensions.rb
@@ -51,20 +51,14 @@ module Danbooru
     end
 
     module ElasticsearchExtensions
-      attr_reader :records_per_page, :total_entries, :current_page, :sequential_paginator_mode
+      include BaseExtension
+
+      attr_reader :total_entries, :current_page, :sequential_paginator_mode
 
       def paginate(page, options = {})
-        @paginator_options = options
+        paginated, mode = paginate_base(page, options)
 
-        validate_page_number(page)
-        sequential = use_sequential_paginator?(page)
-        paginated = if sequential
-                      paginate_sequential(page)
-                    else
-                      paginate_numbered(page)
-                    end
-
-        new_opts = {mode: (sequential ? :sequential : :numbered), seq_mode: sequential_paginator_mode,
+        new_opts = {mode: mode, seq_mode: sequential_paginator_mode,
                     per_page: records_per_page, total: total_count, current_page: current_page}
         if options[:results] == :results
           PaginatedArray.new(paginated.results, new_opts)
@@ -73,30 +67,9 @@ module Danbooru
         end
       end
 
-      def validate_page_number(page)
-        return if page.is_a? Numeric
-        return if page.blank?
-        raise ::Danbooru::Paginator::PaginationError.new("Invalid page number.") unless page =~ /\A[ab]?\d+\z/i
-      end
-
-      def use_sequential_paginator?(page)
-        page =~ /[ab]\d+/i
-      end
-
-      def paginate_sequential(page)
-        if page =~ /b(\d+)/
-          paginate_sequential_before($1)
-        elsif page =~ /a(\d+)/
-          paginate_sequential_after($1)
-        else
-          paginate_sequential_before
-        end
-      end
-
       def paginate_sequential_before(before_id = nil)
         search.definition.update(size: records_per_page + 1, track_total_hits: records_per_page+1)
         search.definition[:body].update(sort: [{id: :desc}])
-
 
         if before_id.to_i > 0
           search.definition[:body][:query][:bool][:must] << ({range: {id: {lt: before_id.to_i}}})
@@ -117,51 +90,16 @@ module Danbooru
       end
 
       def paginate_numbered(page)
-        page = [page.to_i, 1].max
-
-        if page > Danbooru.config.max_numbered_pages
-          raise ::Danbooru::Paginator::PaginationError.new("You cannot go beyond page #{Danbooru.config.max_numbered_pages}. Please narrow your search terms.")
-        end
-
+        page = validate_numbered_page(page)
         search.definition.update(size: records_per_page, from: (page - 1) * records_per_page, track_total_hits: Danbooru.config.max_numbered_pages * records_per_page + 1)
         @current_page = page
 
         self
       end
 
-      def records_per_page
-        option_for(:limit).to_i
-      end
-
       def limit(count)
         search.definition.update(size: count)
         self
-      end
-
-      # When paginating large tables, we want to avoid doing an expensive count query
-      # when the result won't even be used. So when calling paginate you can pass in
-      # an optional :search_count key which points to the search params. If these params
-      # exist, then assume we're doing a search and don't override the default count
-      # behavior. Otherwise, just return some large number so the paginator skips the
-      # count.
-      def option_for(key)
-        case key
-        when :limit
-          limit = @paginator_options.try(:[], :limit) || Danbooru.config.posts_per_page
-          if limit.to_i > 320
-            limit = 320
-          end
-          limit
-
-        when :count
-          if @paginator_options.has_key?(:search_count) && @paginator_options[:search_count].blank?
-            1_000_000
-          elsif @paginator_options[:count]
-            @paginator_options[:count]
-          else
-            nil
-          end
-        end
       end
 
       def total_count
